# docs/swagger.yaml
basePath: /
host: localhost:8080
definitions:
  recipebox-backend-go_internal_dto.AuthEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.AuthResponse'
    type: object
  recipebox-backend-go_internal_dto.AuthResponse:
    properties:
      tokens:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.TokenPair'
      user:
        $ref: '#/definitions/recipebox-backend-go_internal_models.User'
    type: object
...
  recipebox-backend-go_internal_dto.EmailRequest:
    properties:
      email:
        example: alkahfii2018@gmail.com
        type: string
    type: object
...
  recipebox-backend-go_internal_dto.LoginRequest:
    properties:
      email:
        example: alkahfii2018@gmail.com
        type: string
      password:
        example: secret123
        type: string
    type: object
...
  recipebox-backend-go_internal_dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  recipebox-backend-go_internal_dto.OneTimeTokenEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.OneTimeTokenResponse'
      message:
        type: string
    type: object
  recipebox-backend-go_internal_dto.OneTimeTokenResponse:
    properties:
      expiresAt:
        type: string
      token:
        type: string
    type: object
...
  recipebox-backend-go_internal_dto.RefreshRequest:
    properties:
      refreshToken:
        example: refresh-token-sample
        type: string
    type: object
  recipebox-backend-go_internal_dto.RegisterEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.RegisterResponse'
    type: object
  recipebox-backend-go_internal_dto.RegisterRequest:
    properties:
      email:
        example: alkahfii2018@gmail.com
        type: string
      name:
        example: Kahfi Smith
        type: string
      password:
        example: secret123
        type: string
    type: object
  recipebox-backend-go_internal_dto.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/recipebox-backend-go_internal_models.User'
    type: object
  recipebox-backend-go_internal_dto.ResetPasswordRequest:
    properties:
      newPassword:
        example: newSecret123
        type: string
      token:
        example: reset-token-sample
        type: string
    type: object
...
  recipebox-backend-go_internal_dto.TokenEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.TokenPair'
    type: object
  recipebox-backend-go_internal_dto.TokenPair:
    properties:
      accessToken:
        type: string
      accessTokenExpiresAt:
        type: string
      refreshToken:
        type: string
      refreshTokenExpiresAt:
        type: string
    type: object
...
  recipebox-backend-go_internal_dto.UserEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_models.User'
    type: object
  recipebox-backend-go_internal_dto.VerifyEmailRequest:
    properties:
      token:
        example: verify-token-sample
        type: string
    type: object
  recipebox-backend-go_internal_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      emailVerifiedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: API for RecipeBox backend services.
  title: RecipeBox Backend API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password.
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.AuthEnvelope'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token and clear cookie.
      parameters:
      - description: Refresh payload
        in: body
        name: payload
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MessageResponse'
      summary: Logout
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: Return current authenticated user profile.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.UserEnvelope'
      summary: Get current user profile
      tags:
      - Auth
  /api/v1/auth/password/forgot:
    post:
      consumes:
      - application/json
      description: Request reset password flow.
      parameters:
      - description: Email payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.OneTimeTokenEnvelope'
      summary: Request password reset token
      tags:
      - Auth
  /api/v1/auth/password/reset:
    post:
      consumes:
      - application/json
      description: Reset password using one-time token.
      parameters:
      - description: Reset password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MessageResponse'
      summary: Reset password
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token (cookie or request body).
      parameters:
      - description: Refresh payload
        in: body
        name: payload
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.TokenEnvelope'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account.
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.RegisterEnvelope'
      summary: Register user
      tags:
      - Auth
  /api/v1/auth/verify-email/confirm:
    post:
      consumes:
      - application/json
      description: Verify email using one-time token.
      parameters:
      - description: Verify email payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MessageResponse'
      summary: Confirm email verification
      tags:
      - Auth
  /api/v1/auth/verify-email/request:
    post:
      consumes:
      - application/json
      description: Request email verification flow.
      parameters:
      - description: Email payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.OneTimeTokenEnvelope'
      summary: Request email verification token
      tags:
      - Auth
...
schemes:
- http
- https
swagger: "2.0"